examples/xApp/python3/requirements-mcp.txt
fastapi
mcp
pydantic
python-dotenv
requests
uvicorn
a2a-sdk[http-server]
langchain
langgraph
langsmith
redis
